net: lan78xx: Disable TCP Segmentation Offload (TSO)
commit  b8707dc82db19759930751c5a4715b5f432c793a from
https://github.com/raspberrypi/linux.git rpi-6.18.y

TSO seems to be having issues when packets are dropped and the
remote end uses Selective Acknowledge (SACK) to denote that
data is missing. The missing data is never resent, so the
connection eventually stalls.

There is a module parameter of enable_tso added to allow
further debugging without forcing a rebuild of the kernel.

https://github.com/raspberrypi/linux/issues/2449
https://github.com/raspberrypi/linux/issues/2482

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>