media: adv7180: Nasty hack to allow input selection.
commit  811f4d2b28d5785a72540baef9169e203cb0684a from
https://github.com/raspberrypi/linux.git rpi-6.18.y

Whilst the adv7180 driver support s_routing, nothing else
does, and there is a missing lump of framework code to
define the mapping from connectors on a board to the inputs
they represent on the ADV7180.

Add a nasty hack to take a module parameter that is passed in
to s_routing on any call to G_STD, or S_STD (or subdev
g_input_status call).

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>