arm64: dts: fsl: Add Foxconn LS1028A-TSN board
commit 655e33db3d61ee07c93e05d28b49e03d00511fb6 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git

The main board features the NXP LS1028A processor for industrial
communications and a NXP SJA1105S switch cascaded from one of the
LS1028A switch ports.

It has 3 RJ-45 and one SFP port.

There is a daughter board that can be plugged into it, that extends the
Ethernet connectivity further with 4 more RJ-45 ports.

Signed-off-by: Wes Li <wes.li@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Jerry Huang <jerry.huang@nxp.com>
Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>