drm/vc4_hdmi: Allow hotplug detect to be forced
commit  0c3a90f3695fbe2e1bca9e54e0beb52d0e858d8c from
https://github.com/raspberrypi/linux.git rpi-6.18.y

See: https://forum.libreelec.tv/thread/24783-tv-avr-turns-back-on-right-after-turning-them-off

While the kernel provides a :D flag for assuming device is connected,
it doesn't stop this function from being called and generating a cec_phys_addr_invalidate
message when hotplug is deasserted.

That message provokes a flurry of CEC messages which for many users results in the TV
switching back on again and it's very hard to get it to stay switched off.

It seems to only occur with an AVR and TV connected but has been observed across a
number of manufacturers.

The issue started with https://github.com/raspberrypi/linux/pull/4371
and this provides an optional way of getting back the old behaviour

Signed-off-by: Dom Cobley <popcornmix@gmail.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>