dmaengine: bcm2835: HACK: Support DMA-Lite channels
commit  32eb163258d055ca52855b6059493bc3f982703a from
https://github.com/raspberrypi/linux.git rpi-6.18.y

The BCM2712 has a DMA-Lite controller that is basically a BCM2835-style
DMA controller that supports 40 bits DMA addresses.

We need it for HDMI audio to work, but this breaks BCM2835-38 so we
should rework this later.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>

dmaengine: bcm2835: Fix dma driver for BCM2835-38

The previous commit broke support on older devices.
Make the breaking parts of patch conditional on
the device being used.

Fixes: 6e1856ac7c39 ("dmaengine: bcm2835: HACK: Support DMA-Lite channels")

Signed-off-by: Dom Cobley <popcornmix@gmail.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>