drivers: char: add generic gpiomem driver
commit  11081dbe94d593ac8685c22afc5a96f635edf8a8 from
https://github.com/raspberrypi/linux.git rpi-6.18.y

Based on bcm2835-gpiomem.

We allow export of the "GPIO registers" to userspace via a chardev as
this allows for finer access control (e.g. users must be group gpio, root
not required).

This driver allows access to either rp1-gpiomem or gpiomem, depending on
which nodes are populated in devicetree.

RP1 has a different look-and-feel to BCM283x SoCs as it has split ranges
for IO controls and the parallel registered OE/IN/OUT access. To handle
this, the driver concatenates the ranges for an IO bank and the
corresponding RIO instance into a contiguous buffer.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>