net: dsa: sja1105: make hostprio configurable through the dtb
commit 26e984722ee785e9ebbfc940e75a598e438e3077 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git

Currently, the default hostprio (management queue) is 7:
which means priority queue used for both ingress and egress for managmenet frames is the highest priority queue.
In AVB usescases, the two highest priority queues are assigned to SR class traffic and management traffic can be assigned to queue 5 for example.
So make this parameter configurable through the device tree.

Signed-off-by: Alexandre Howard <alexandre.howard@nxp.com>
Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>