net: stmmac: fix carrier check handling in stmmac_xdp_xmit_zc()
commit 1b2d1acdc729362d506afcd7c2322331365b61e9 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git

When using AF_XDP Zero Copy, a cable unplug will make the napi
kernel thread (when using threaded napi) consumes 100% of CPU load
as a result of a continuous napi reschedule, since stmmac_xdp_xmit_zc()
states that there is still work to be done.
The fix is to return work_done as true when carrier is not ok.
A similar fix was provided for ixgbe driver:
"6c7273a2667 ixgbe: xsk: change !netif_carrier_ok() handling in ixgbe_xmit_zc()"

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>
Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>