net: sdk_dpaa: allow RX checksum offload bypass using NETIF_F_RXCSUM
commit f62644bdcb4577b8b379440073b5f403b4c67080 from
https://github.com/nxp-imx/linux-imx.git

L4 RX checksum validation is performed by default whenever the parser is
enabled, aka when PCDs are applied. We should report the feature to
ethtool.

Additionally, through the DCL4C bit of the FMBM_RFNE[FDCS] register,
we can force the FMan to treat packets as having a good RX checksum.
Some customers want to be able to receive packets with RX checksum
errors. The way this would be possible is to set that bit.

This is exposed by adding a FMan wrapper function to export the API
that configures the register. Then we call it when NETIF_F_RXCSUM gets
disabled.

When FMBM_RFNE[FDCS] is set, we still want packets with invalid RX
checksum to be dropped by default, so checksumming needs to happen in
software. For that, make sure we don't set CHECKSUM_UNNECESSARY, as that
would be incorrect. Just leave them at CHECKSUM_NONE.

Some further modifications will be necessary to be able to actually
process packets with wrong checksum instead of having the network stack
drop them, like the SO_NO_CHECK socket option for instance.

Co-developed-by: Camelia Groza <camelia.groza@nxp.com>
Signed-off-by: Camelia Groza <camelia.groza@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>