media: amphion: Abort vpu parsing directly in seek
commit 168cffba435fa3c2505ece5aa485346112fd5f4b from
https://github.com/nxp-imx/linux-imx.git

Driver abort vpu decoding when both output and capture queues are off,
but if seek in parsing the sequence header, driver may miss aborting the
parsing. so just abort the vpu parsing directly in seek.

Meanwhile if capture is off unexpectedly, we still need to abort the
decoding and return capture buffers.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>