phy: ti: add PHY driver for TI DS125DF111 Dual-Channel Retimer
commit 5e11d11095efcd1d001c4a542a0dfdca1cec11c4 from
https://github.com/nxp-imx/linux-imx.git

Add a generic PHY driver for the TI DS125DF111 Multi-Protocol
Dual-Channel Retimer. The driver currently supports only 10G and 1G link
speeds but it can easily extended to also cover other usecases.

Since the available datasheet (https://www.ti.com/lit/gpn/DS125DF111)
does not name the registers, the driver uses direct values for them
instead of macros.

A PHY device is created for each of the two channels present on the
retimer.  This allows for independent configuration of the two channels.
This capability is especially important on retimers which have more than
2 channels that can be, depending on the board design, connected in
multiple different ways to the SerDes lanes.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>