drivers: gpio: qixis-cpld: extend support also for the LS1046AQDS QIXIS CPLD
commit 34bbf282e31e6c99c6eeda742004c231e8a395da from
https://github.com/nxp-imx/linux-imx.git

Extend the driver to also support the GPIO controller found on the
LS1046AQDS QIXIS CPLD.

For usecases in which the parent device (the QIXIS CPLD) does not define
a regmap for its memory area, create the driver's own regmap which will
get passed to gpio-regmap trough devm_gpio_regmap_register().
In this case, the offset inside of the regmap will be zero since the
regmap points directly to the desired register.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>