arm64: dts: ls1046a-qds: describe the two on-board SFP+ cages
commit b16663d72f6ae10f20630e77504c11f0ebf4ec56 from
https://github.com/nxp-imx/linux-imx.git

Describe the two SFP+ cages present on the LS1046AQDS board and their
associated I2C buses and GPIO lines.

The QIXIS CPLD node is also extended so that it describes the GPIO
controller responsible for all the status presence lines on both SFP+
cages as well as the IO SLOTs present on the board.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>