i2c: mux: Add support for generic base-nr property
commit  6aeca2833f5a3dcf402d359e800910ce1776c6e4 from
https://github.com/raspberrypi/linux.git rpi-6.18.y

i2c_mux_add_adapter takes a force_nr parameter that allows an explicit
bus number to be associated with a channel. However, only i2c-mux-reg
and i2c-mux-gpio make use of it.

To help with situations where it is desirable to have a fixed, known
base address for the channels of a mux, create a "base-nr" property.
When force_nr is 0 and base-nr is set and non-zero, form a force_nr
value from the sum of base-nr and the channel ID.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>