LF-11724-2: media: os08a20: Implement V4L2_CID_HDR_SENSOR_MODE
commit 1761777d6742023f55cc7b8f371fb8029edb6cc6 from
https://github.com/nxp-imx/linux-imx.git

The information from S_FMT (resolution and pixel code) is not
enough to define the sensor mode, further more, we need to know
the hdr mode. For that, implement V4L2_CID_HDR_SENSOR_MODE.
We need to take into account that s_ctrl and S_FMT may come
in any order. So let the S_FMT establish the resolution
and pixel code and use whatever hdr mode was previously set,
while the s_ctrl will try to find a mode that matches the
current resolution and code and the new hdr mode.
Both S_FMT and s_ctrl will update the sensor if needed and
if a matching mode is found.

Signed-off-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Alice Yuan <alice.yuan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>