LF-11845-2 net: enetc: add link status message support for i.MX95
commit 01d489d28ad0539d97129e3e063b2f18c5d4aacc from
https://github.com/nxp-imx/linux-imx.git

The VF should have the ability to obtain the current link status of the
PF in order to decide whether to set or clear the carrier. Otherwise,
even if the network cable is not plugged in, VF will always deem that
its network is connected, which is very confusing.
So we utilize the mailbox feature of ENETC to transmit the current link
status of PF to the VF, so that the VF has the ability to detect the
current link status of PF.
The basic test results are as expected and are shown below.
1. Disable or enable the interface of PF, the VF interfaces can detect
the link change.
root@imx95-19x19-lpddr5-evk:~# ifconfig eth0 down
[   41.912555] fsl_enetc_vf 0002:00:04.0 eth3: Link is Down
[   41.912564] fsl_enetc_vf 0002:00:02.0 eth2: Link is Down
[   41.912618] fsl_enetc4 0002:00:00.0 eth0: Link is Down
root@imx95-19x19-lpddr5-evk:~# ifconfig eth0 up
[   59.249284] fsl_enetc4 0002:00:00.0 eth0: PHY [0003:01:00.0:01] driver [RTL8211F-VD Gigabit Ethernet] (irq=POLL)
[   59.260576] fsl_enetc4 0002:00:00.0 eth0: configuring for phy/rgmii-id link mode
[   62.374439] fsl_enetc_vf 0002:00:02.0 eth2: Link is Up
[   62.374447] fsl_enetc_vf 0002:00:04.0 eth3: Link is Up
[   62.374526] fsl_enetc4 0002:00:00.0 eth0: Link is Up - 1Gbps/Full - flow control rx/tx

2. Unplug and replug the network cable, the VF interfaces can also
detect the link change.
[   85.926190] fsl_enetc_vf 0002:00:02.0 eth2: Link is Down
[   85.926204] fsl_enetc_vf 0002:00:04.0 eth3: Link is Down
[   85.926271] fsl_enetc4 0002:00:00.0 eth0: Link is Down
[   90.027075] fsl_enetc_vf 0002:00:04.0 eth3: Link is Up
[   90.027077] fsl_enetc_vf 0002:00:02.0 eth2: Link is Up
[   90.027161] fsl_enetc4 0002:00:00.0 eth0: Link is Up - 1Gbps/Full - flow control rx/tx

3. Disable and then enable the VF interface.
root@imx95-19x19-lpddr5-evk:~# ifconfig eth2 down
root@imx95-19x19-lpddr5-evk:~# ifconfig eth2 up
[  735.023908] fsl_enetc_vf 0002:00:02.0 eth2: Link is Up

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>