staging: fsl_qbman: update PVB bit during drain_mr_fqrni()
commit 6b669bd3bf7a4caea6c7681408ccafc9af88e97e from
https://github.com/nxp-imx/linux-imx.git

If FQRNI events are found in the message ring and we call
drain_mr_fqrni(), things don't work because the code doesn't update
the alternating Polarity Valid Bit as it should, when walking the MR.

Similar to upstream commit 627da8bad530 ("soc/fsl/qbman: Fix
drain_mr_fqni()"). This was fixed upstream but not for the SDK QBMan
driver.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>