staging: fsl_qbman: drain FQ after retirement with a single VDQCR command
commit b090229378611753b7521b5585fce29b249a6bdc from
https://github.com/nxp-imx/linux-imx.git

qm_shutdown_fq() open-codes qman_volatile_dequeue(), but as opposed to
that, it does not set p->vdqcr_owned for the portal.

A race condition can be observed, when draining the FQ under high
traffic, that sometimes after the QMan interrupts are re-enabled,
a QM_DQRR_STAT_FQ_EMPTY event is received on this frame queue in
portal_isr() -> __poll_portal_fast(), but we aren't prepared to
handle it.

Use the QM_VDQCR_NUMFRAMES_TILLEMPTY option when issuing the volatile
dequeue command, and wait inline until it finishes by signaling to us
the QM_DQRR_STAT_FQ_EMPTY event. This way, this interrupt will not be
pending after the FQ shutdown is finished.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>