staging: fsl_qbman: consolidate STATE_RETIRED processing with the others in qm_shutdown_fq()
commit 46b3e2e8cb4398710d2191fbcaf7f1125a528ad0 from
https://github.com/nxp-imx/linux-imx.git

The last bit of code for the STATE_TEN_SCHED, STATE_TRU_SCHED, STATE_ACTIVE
and STATE_PARKED states is duplicated for the STATE_RETIRED state. That is,
bringing the FQ out of service.

Use a "fallthrough" statement to avoid code duplication and possible
bugs caused by having different implementations (as the one that was
just fixed by the previous commit).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>