arm64: dts: ls1046a: add aliases for QMan and BMan reserved-memory nodes
commit a5fa17c1c2b73bc30f683f55277ea1e21e48c273 from
https://github.com/nxp-imx/linux-imx.git

For kexec to reliably work, we need all kernels to see the same memory
region base addresses for the QMan PFDR and FQD and the BMan FBPR. This
is because we don't reset and can't reinitialize these peripherals (and
these reserved-memory regions are write-once).

The current reserved-memory format we use for these nodes is not good.
It requests a dynamic allocation, which means it's possible for one
kernel to allocate a different reserved-memory region than the previous
one. What we need are static allocations. But.. we can't just statically
hardcode those in the ls1046a.dtsi or even in board device trees,
because it depends on the quantity of available RAM. So we made newer
U-Boot versions to carve out memory from the top of RAM and fix up the
device tree.

The problem is that we don't want to hardcode the node names in U-Boot
because that's frowned upon (for good reason, as it caused breakage
before):
https://lore.kernel.org/lkml/20210612075346.GZ29138@dragon/

So we need aliases for U-Boot to get to these reserved-memory regions.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>