LF-11321-2 net: enetc: make PF driver only provide SRIOV interface
commit b66e311a4d43e850eb26f96de6f6bc824d12c98c from
https://github.com/nxp-imx/linux-imx.git

As we know, the PCI PF driver provides SRIOV interface to enable
its own VFs. But there is an use case that PF is owned by M core
and VFs are owned by A core. In this case the PF driver in Linux
kernel should only provide SRIOV interface so that we can enable
VFs through the standard interface and command. For probe() and
remove(), success is returned directly. Then we can use the below
command to enable VFs.

echo 2 > /sys/bus/pci/devices/0002\:00\:10.0/sriov_numvfs

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Reviewed-by: Frank Li <frank.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>