net: dpaa2-eth: add MACSec SecY offload support
commit 6b44fd639f1be3ac3654f428e04ce5c303fd31cb from
https://github.com/nxp-imx/linux-imx.git

This patch adds support for offloading MACSec SecY instances.

First of all, the necessary MC firmware APIs for creating and
configuring a secure instance are added.
The dpni_is_macsec_capable() API was added in order to be able to verify
if the current layout of DPAA2 objects enables a DPNI to access the
MACSec feature ie. is connected to a MACSec capable DPMAC.

The next step is to implement the SecY relevant macsec_ops callbacks in
order to propagate each userspace request for add, removing or
configuring a SecY to the HW.

It's important to note the fact that the dpaa2-eth driver will
restrict the number of SecYs that can be created on a interface to only
one. This is because we do not have available metadata in each frame to
determine which Secure Channel a specific MACSec frame came from, the
only thing that we know is the fact that it is a MACSec frame.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>