LF-12071 gpio: vf610: always set GPIO to input mode when used as interrupt source
commit 9629edfd833c1fbcb795132b28e1aad4dcce2474 from
https://github.com/nxp-imx/linux-imx.git

Though the default pin configuration is INPUT, but if the prior stage does
configure the pins as OUTPUT, then Linux will not reconfigure the pin as
INPUT. If use this pin as interrupt source, and set as low level trigger,
if prior stage set this pin as OUTPUT low, then will meet interrupt storm.

So always set GPIO to input mode when used as interrupt source to fix above
case.

Tested-by: Xu Yang <xu.yang_2@nxp.com>
Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>