LF-11880-2 net: enetc: Add power off suspend/resume and WOL supports
commit 1b34970b6c374f99c7aa5fdc153cc157622ba1a4 from
https://github.com/nxp-imx/linux-imx.git

In previous implementations, NETCMIX does not power off during suspend.
If NETCMIX powers off, all NETCMIX devices need to be reinitialized in the
resume process.
So, add this function support.

Support Wake-on-LAN functionality in suspend at the same time.
Need to execute the following command to set it as WOL mode.
 # ethtool -s ethX wol g

When using WOL mode, NETCMIX will not power off during suspend.

Reviewed-by: Claudiu Manoil <claudiu.manoil@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>