LF-11880-2 net: enetc: Add power off suspend/resume and WOL supports
commit 438c5e9014853d1bb0b22e20518d9c6eb02601a5 from
https://github.com/nxp-imx/linux-imx.git

In previous implementations, NETCMIX does not power off during suspend.
If NETCMIX powers off, all NETCMIX devices need to be reinitialized in the
resume process.
So, add this function support.

Support Wake-on-LAN functionality in suspend at the same time.
Need to execute the following command to set it as WOL mode.
 # ethtool -s ethX wol g

When using WOL mode, NETCMIX will not power off during suspend.

Reviewed-by: Claudiu Manoil <claudiu.manoil@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>