LF-12108-1 net: enetc: ierb: allocate MSI interrupt LDIDs for the two VFs of port1
commit d7c9bc1b18529d9a4f9e4cfab0dc3ac9e5e540f0 from
https://github.com/nxp-imx/linux-imx.git

IMX95 only provided 8 MSI LDIDs, but there are more than 8 PCIe devices
requiring interrupts. In 19x19, we disabled the VF function for port1 to
support all PF and VF functions of port0 and port2. In 15x15, port2 is no
longer available, so the interrupts previously used by its VFs are now
allocated to the VFs of port1.

Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>