drm/vc4: Fixup mode for 7inch panel on DSI0
commit  62ef117360d33c336c125acdff7cf585b2a3f7f8 from
https://github.com/raspberrypi/linux.git rpi-6.18.y

The TC358762 bridge and panel decodes the mode differently on
DSI0 to DSI1 for no obvious reason, and results in a shift off
the screen.
Whilst it would be possible to change the compatible used for
the panel, that then messes up Pi5.

As it appears to be restricted to vc4 DSI0, fix up the mode
in vc4_dsi.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>