LF-10512: media: imx: ci_pi: fix green line at frame bottom issue
commit c2596b799d49c4613cb97fdab13904201d6fd144 from
https://github.com/nxp-imx/linux-imx.git

The bottom line of image will be green since the data are zero.
The reason is due to vsync signal polarity mismatch between camera
and CI_PI expected.

The CI_PI expected vsync signal like this:
        ┌─┐                      ┌─┐
	│ │                      │ │
VSYNC ──┘ └──────────────────────┘ └──

The vsync signal of camera used on iMX93/91 like this:

VSYNC ──┐ ┌──────────────────────┐ ┌──
        │ │                      │ │
	└─┘                      └─┘

So need to invert the vsync signal polarity in CI_PI side.

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>