LF-12371 net: enetc: fix suspend/resume issue when eth port down
commit a792dd64a841963e3c24894674026fa031e00b5b from
https://github.com/nxp-imx/linux-imx.git

When an Ethernet port is down, the resume function will not re-initialize
the PCI-related registers of the Ethernet port. However, at this point,
the port has already been power-cycled, causing the port to be unusable
when up after the resumen.

To fix this, during suspend/resume, even if the port is not running,
these registers will be initialized as well.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>