arm64: dts: fix avb timers audio pll input clock frequency
commit 47255e41d56910075c942575eb3452f55b8ceaaf from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git

Limit the audio pll frequency input for the AVB timers to the maximum
allowed for the timer on each SoC (specified in their respective Reference Manuals).
For i.MX 8MPlus and i.MX 8MMini, the maximum frequency for GPT is 100 MHz.
For i.MX 93, the maximum frequency for TPM is 83 MHz

fixes: "6318d349398 arm64: dts: set avb timers to the audio pll rate"

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>