LF-12343 arch: arm64: ixm8mp-evk-sof-wm8960: Fix codec MLCK rate setting race
commit 8180af35713c7661f31375b763dd2c62e52cc53c from
https://github.com/nxp-imx/linux-imx.git

Current implementation for i.MX8MP with wm8960 sets the clock hierarchy
and rates for SAI in the dsp node.

This is problematic because sometimes the clock reparenting / rate set
can happen after wm8960 I2C driver probes.

wm8960 I2C reads mclk rate in the probe function (and uses it later).
This value is outdated.

Sequence of events:

Core 0                                                          Core 1
-----                                                         ------
wm8960_i2c_probe
 -> wm8960->mclk = devm_clk_get(&i2c->dev, "mclk");
 -> wm8960->freq_in = clk_get_rate(wm8960->mclk);
 -> Here we get wm8960->freq_in = 24000000
                                                              snd_sof_of_imx8m_driver_init
                                                              -> of_clk_set_defaults
                                                                -> here set wm8960->mclk 12288000
wm8960_hw_params
  -> wm8960_configure_clocking
     -> configure BCLK and LRCLK based on old wm8960->freq

As you can see we set BLCK and LRCLK based on a wrong MCLK so the
rate playback goes is wrong.

This problem doesn't happen for wm8962 codec because BLCK/LRCLK is
configured via sysclk and that's after probe when the clock hierarchy
is already set.

The problem doesn't happen from wm8960 with imx8qxp/imx8qm because on
those platforms MCLK is set inside the wm8960 device tree node and at
the point of the probe the clock hierarchy is already known.

We take a similar approach as  for imx8qxp and imx8qm and move the
parent/rate setup in codec dts node.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
Reviewed-by: Iuliana Prodan <iuliana.prodan@nxp.com>
Reviewed-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>