MLK-26186: media: os08a20: Fix V4L2_CID_ANALOGUE_GAIN
commit 7b8a5b629f541f3ed31ca50ebeb29c04a2e8a6c1 from
https://github.com/nxp-imx/linux-imx.git

The os08a20_set_gain() function was inherited from imx8mp,
and it is specific to the VSI ISP, which is expecting a total
gain of 1024 for 1x gain, with the analogue and digital gain
being combined.

For imx95, add a new function and let the control update only
the analog gain registers with the gain code in sensor format.
This covers linear mode (gain for long exposure).

Fixes: 4892da9340be ("MLK-26186-5: media: os08a20: Add analog gain and exposure controls")

Signed-off-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>