LF-12679-2 media/i2c: mx95mbcam: Reset serializer before removal or reboot
commit fb6e98560e34e127729f8760df74af4e5777f731 from
https://github.com/nxp-imx/linux-imx.git

Since the MX95MBCAM sensor module has the PWDNB pin, that should reset
the serializer, hard-wired to HIGH, we need to make sure we leave the
serializer in the default state when we remove the module or we perform
a warm reboot. This would avoid issues with GMSL links not locking when
rebooting the board.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>