LF-12679-3 media/i2c: max96724: Add a more robust GMSL link check mechanism
commit c0af0cfd114cdf3093c0f360da517f6f4195ddf7 from
https://github.com/nxp-imx/linux-imx.git

It may happen when booting the system up that GMSL links do not lock.
However, a simple chip reset can fix it easily. So, this patch modifies
the init sequence by implementing a retry mechanism that checks the GMSL
links, resetting the chip in case no link has locked. The checks will be
done at the beginning of the init sequence, instead of the end.  Also,
remove the links reset call from max96724_reset_gmsl_links() and rename
the function to max96724_check_gmsl_links().

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>