AIR-11581 driver: neutron: Enable runtime power-management autosuspend
commit ac3817553c12277924ba3f552893e331f643b159 from
https://github.com/nxp-imx/linux-imx.git

When the Neutron NPU (Neural Processing Unit) is in idle status and not
doing inference tasks, it is desirable for the system to automatically
move NPU into suspend state. This behavior could reduce power consumption
for the whole Soc.

This patch is to update the Neutron driver by integrating autosuspend
and resume capabilities. This update allows the driver to intelligently
manage the power state of the NPU, automatically placing it into a
suspend status when it detects an extended period of inactivity.

Signed-off-by: Jiwei.Fu <jiwei.fu@nxp.com>
Reviewed-by: Forrest Shi <xuelin.shi@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>