LF-12898-1: Touchscreen: Add reset-gpio support in focaltech touchscreen driver
commit 11a9a695c701b4930fb3bad02244d7d7c1df5b61 from
https://github.com/nxp-imx/linux-imx.git

Touchscreen need a reset signal to wake up FT3518 from the Hibernate mode.
After resetting, FT3518 shall enter the Active mode. Then I2C IO operation
to FT3518 will take effect. So add the reset-gpio support in the focaltech
touchscreen driver.

Some pannels are designed to share the same reset pin for display pannel
and touchscreen, and the reset pin is controlled by the display. In this
case, display pannel driver must be probed before the touchscreen driver
and should send a reset signal, so the touchscreen can be enabled through
the reset signal of the display. But when display and touch each have an
independent reset pin, the reset pin of the touchscreen must be enabled.

Signed-off-by: Carlos Song <carlos.song@nxp.com>
Reviewed-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>