LF-12919 net: enetc: fix BD ring interruption affinity issue
commit 1b34970b6c374f99c7aa5fdc153cc157622ba1a4 from
https://github.com/nxp-imx/linux-imx.git

When system suspend, the interruption affinity settings become
ineffective due to the shutdown of multi-core CPUs.
Enabling Wake-on-LAN (WoL) will result in interruption affinity
being bound to CPU0 for all bdrings due to the lack of
reinitialization of interrupts during the suspend/resume process,
as NETCMIX remains powered.

The function enetc_restore_irqs_affinity() is added to restore
the interruption affinity settings.

Fixes: 438c5e901485 ("LF-11880-2 net: enetc: Add power off suspend/resume and WOL supports")

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>