net: pcs: mtip_backplane: make mtip_are_all_lanes_trained() return true for 1000Base-KX
commit 5d32469ce01f855f508eee1a0c6be38c708b0bb4 from
https://github.com/nxp-imx/linux-imx.git

Currently the link is held down by mtip_backplane_get_state() when C73
resolves to 1000Base-KX, because this mode does not require link
training, and the function is coded up to return false to that question.

Allow to link to come up immediately after auto-negotiation resolves
1000Base-KX, by instead responding "yes, all lanes are trained".

priv->link_mode is only valid when priv->link_mode_resolved is true,
hence the two checks. On the other hand, the priv->link_mode_resolved
check doesn't impact the existing logic, because link training comes
after the link mode was resolved (so it must be true in order for all
lanes to have been trained).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>