net: dpaa2: allow dpmac_set_protocol("1000base-kx") to succeed
commit 4b5b0ae3dc24b620dae585277b950b884c5fa36a from
https://github.com/nxp-imx/linux-imx.git

Map PHY_INTERFACE_MODE_1000BASEKX to a reasonable DPMAC_ETH_IF_* value
for 1G to pass to the MC firmware, like DPMAC_ETH_IF_1000BASEX. It is
necessary in order to reconfigure the MAC if coming from a higher speed.
BASEX/SGMII/BASE-KX does not actually matter, since all of that is below
the MAC layer and is handled by Linux.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>