net: mscc: ocelot: do not use ocelot_port->phy_mode in ocelot_phylink_mac_config()
commit 252cd4b7de1866189557a9f9cb70f4397a2a9ad1 from
https://github.com/nxp-imx/linux-imx.git

phylink_mac_ops :: mac_config() already provides a valid state->interface
which should be equal to the ocelot_port->phy_mode. But ocelot_port->phy_mode
is fixed, whereas state->interface can be dynamic if we allow it to.

Use what phylink gives us, not state cached in the driver.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>