drm/vc4: Use the TPZ scaling filter for 1x1 source images
commit  198ef986b23cc237890f4072cc59b14f216e88ee from
https://github.com/raspberrypi/linux.git rpi-6.18.y

The documentation says that the TPZ filter can not upscale,
and requesting a scaling factor > 1:1 will output the original
image in the top left, and repeat the right/bottom most pixels
thereafter.
That fits perfectly with upscaling a 1x1 image which is done
a fair amount by some compositors to give solid colour, and it
saves a large amount of LBM (TPZ is based on src size, whilst
PPF is based on dest size).

Select TPZ filter for images with source rectangle <=1.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>