dmaengine: dw-axi-dmac: Allow client-chosen width
commit  615e97109e9bf453d783273928dd92974f03768c from
https://github.com/raspberrypi/linux.git rpi-6.18.y

For devices where transfer lengths are not known upfront, there is a
danger when the destination is wider than the source that partial words
can be lost at the end of a transfer. Ideally the controller would be
able to flush the residue, but it can't - it's not even possible to tell
that there is any.

Instead, allow the client driver to avoid the problem by setting a
smaller width.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>