NotForUpstream: media: video-mux: Propagate controls to source
commit  c5b3d942ccbfda8db53e192b6a7537c3c42fd74f from
https://github.com/raspberrypi/linux.git rpi-6.18.y

The i.MX8MP makes calls on it's source device to determine
the link-frequency that should be configured on the CSI2 receiver.

When the source is behind a video mux, we need to pass this call through
to the connected device.

Map the control handler of the source device to the video-mux,
essentially proxying all controls on the mux to the device which has
it's link enabled.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>