drm/vc4: hvs: Defer updating the enable_bg_fill until vblank
commit  d43e5d64b3a4575114d797258309ffdbf1f18acb from
https://github.com/raspberrypi/linux.git rpi-6.18.y

The register to enable/disable background fill was being set
from atomic flush, however that will be applied immediately and
can be a while before the vblank. If it was required for the
current frame but not for the next one, that can result in
corruption for part of the current frame.

Store the state in vc4_hvs, and update it on vblank.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>