misc: Add ws2812-pio-rp1 driver
commit  949e96a52dacd48c86b887af6fd169f28ca40771 from
https://github.com/raspberrypi/linux.git rpi-6.18.y

ws2812-pio-rp1 is a PIO-based driver for WS2812 LEDS. It creates a
character device in /dev, the default name of which is /dev/leds<n>,
where <n> is the instance number. The number of LEDS should be set
in the DT overlay, as should whether it is RGB or RGBW, and the default
brightness.

Write data to the /dev/* entry in a 4 bytes-per-pixel format in RGBW
order:

  RR GG BB WW RR GG BB WW ...

The white values are ignored unless the rgbw flag is set for the device.

To change the brightness, write a single byte to offset 0, 255 being
full brightness and 0 being off.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>