mmc: sd: filter card CQ support based on an allow-list
commit  67fe44d8f10c3ad29d59818c10e9667dfac46cf2 from
https://github.com/raspberrypi/linux.git rpi-6.18.y

We have found that many SD cards in the field, even of the same make and
model, have latent bugs in their CQ implementation. Some product lines
have fewer bugs with newer manufacture dates, but this is not a
guarantee that a particular card is at a particular firmware revision
level.

Many of these bugs lead to card hangs or data corruption. Add a quirk to
mark a card as having a tested, working CQ implementation and ignore the
capability if absent.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>