drm/vc4: PV1 can be driven via any HVS channel, so adapt to allow it
commit  777b4044a485e2bf6bd7079a072a54c5997911f9 from
https://github.com/raspberrypi/linux.git rpi-6.18.y

It was noted that if PV1 was in use to drive DSI1, then the
writeback connector could not be used as HVS channel 2 was
already in use.
The HVS allows PV1 (HVS output 2) to be driven by any HVS
channel via the DSP3_MUX setting, but that was hardcoded to be
either 2 (for PV1) or disabled for TXP.

Expand the available channels field for PV1, and configure
DSP3_MUX accordingly.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>