drm/vc4: tests: Update pv muxing tests now DSI1 is more flexible
commit  9e11105c7a24ccf18f2d4b3bed79027383fac5be from
https://github.com/raspberrypi/linux.git rpi-6.18.y

The tests on vc4 (BCM2835-7) were checking for DSI1 muxing being
to restricted channel 2, and therefore muxing with TXP was impossible.

As we no longer have that restriction, update the capabilities
defined for DSI1, move the tests that used to be impossible to the
valid list, and extend for additional combinations that are now
possible.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>